Added scrollwheel control and input field focus checks
Also normalized onscreen control colours.

Input field focus check works for search field and for the comment field
via focus and blur event listeners. The event listener was added to a
containing div with useCapture set to true for the comments field as the
input is set to type="hidden".